backend/requirements.txt
Django==2.2.16
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
isort==5.12.0
mccabe==0.7.0
pep8-naming==0.13.3
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.7.1
sqlparse==0.4.3
djangorestframework==3.12.4
Pillow==9.4.0
django-filter==22.1
